add changes
1. откорректировал пагинацию
